Ajout de collisions et thread des balles

+++ src/ballType.hpp
@@ -8,14 +8,14 @@ template <typename FontT>
 class BallType {
     private:
         const size_t DEGATS;
-        const size_t RAYON;
+        const float RAYON;
         const FontT FONT_BASE;
     
     public:
         BallType() = delete;
-        inline BallType(FontT font, size_t rayon, size_t degats): DEGATS{degats}, RAYON{rayon}, FONT_BASE{font} {};
+        inline BallType(FontT font, float rayon, size_t degats): DEGATS{degats}, RAYON{rayon}, FONT_BASE{font} {};
 
-        inline size_t getRayon() const {return RAYON;};
+        inline float getRayon() const {return RAYON;};
         inline size_t getDegats() const {return DEGATS;};
         inline FontT getFont() const {return FONT_BASE;};
 };
